results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/9/LLMatic/recommendation.py
class Recommendation:
	def __init__(self, user):
		self.user = user
		self.tips = []
		self.financial_products = []
		self.generate_tips()
		self.recommend_products()

	def generate_tips(self):
		# This is a placeholder. In a real-world application, this method would analyze the user's financial data and generate personalized tips.
		if 'Save more money' not in self.tips:
			self.tips.append('Save more money')

	def recommend_products(self):
		# This is a placeholder. In a real-world application, this method would analyze the user's financial data and recommend suitable financial products.
		if 'Investment product' not in self.financial_products:
			self.financial_products.append('Investment product')

class RecommendationManager:
	def __init__(self):
		self.recommendations = {}

	def create_recommendation(self, user):
		recommendation = Recommendation(user)
		self.recommendations[user.username] = recommendation
		return recommendation

	def get_recommendation(self, username):
		return self.recommendations.get(username)
