results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/9/LLMatic/test_recommendation.py
import pytest
from recommendation import Recommendation, RecommendationManager
from user import User


def test_recommendation():
	user = User('John Doe', 'password', 'john.doe@example.com')
	recommendation = Recommendation(user)
	recommendation.generate_tips()
	recommendation.recommend_products()
	assert len(recommendation.tips) == 1
	assert recommendation.tips[0] == 'Save more money'
	assert recommendation.financial_products == ['Investment product']


def test_recommendation_manager():
	manager = RecommendationManager()
	user = User('John Doe', 'password', 'john.doe@example.com')
	recommendation = manager.create_recommendation(user)
	assert recommendation == manager.get_recommendation(user.username)
	assert recommendation.tips == ['Save more money']
	assert recommendation.financial_products == ['Investment product']
